requirements.txt
arrow==1.2.2
asciimatics==1.14.0
attrs==21.4.0
aws-requests-auth==0.4.3
AWSIoTPythonSDK==1.5.2
backports.entry-points-selectable==1.1.1
bitstring==3.1.9
bleach==5.0.0
boto3==1.22.13
botocore==1.25.13
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
coverage==6.3.2
cryptography==37.0.2
decorator==5.1.1
distlib==0.3.4
dnslib==0.9.19
dnspython==2.2.1
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.18.1
ecdsa==0.17.0
esptool==4.1
filelock==3.6.0
future==0.18.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
invoke==1.7.1
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.5.1
keyring==23.5.0
MarkupSafe==2.1.1
packaging==21.3
paho-mqtt==1.6.1
pbr==5.9.0
pexpect==4.8.0
Pillow==9.1.0
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
prettytable==3.3.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
pycparser==2.21
pyfiglet==0.8.post1
Pygments==2.12.0
PyJWT==2.4.0
pykwalify==1.8.0
pyparsing==3.0.9
pypng==0.0.21
pyrsistent==0.18.1
pyserial==3.5
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
python-box==6.0.2
python-dateutil==2.8.2
pyusb==1.2.1
PyYAML==6.0
questionary==1.10.0
readme-renderer==35.0
reedsolo==1.5.4
requests==2.27.1
requests-aws4auth==1.1.2
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
semver==2.13.0
six==1.16.0
sqlitedict==2.0.0
stevedore==3.5.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
twine==4.0.0
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.19.0
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
zipp==3.8.0
